Add X-UA-Compatible meta tag
Internet Explorer 11 sometimes mistakenly thinks it is a good idea
to render the site in document mode 7 (because of some settings we
have no control over). This causes all kind of problems with our
grid view setup.
When adding the following meta tag to the HTML of the page we can
force IE11 to render in document mode 11.
<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">